+++ lib/class_functions.php
@@ -1426,44 +1426,6 @@ class WPP_F extends UsabilityDynamics\Utility
   }
 
   /**
-   * Default property page url
-   *
-   * author Den@UD
-   */
-  static public function register_property_page_url()
-  {
-    global $wpdb, $wp_properties;
-
-    $base_slug = $wp_properties['configuration']['base_slug'];
-
-    //** Check if this page actually exists */
-    $post_id = $wpdb->get_var("SELECT ID FROM {$wpdb->posts} WHERE post_name = '{$base_slug}'");
-
-    $post_url = get_permalink($post_id);
-
-    $wp_properties['configuration']['base_property_url'] = $post_url;
-    update_option('wpp_settings', $wp_properties);
-
-  }
-
-  /**
-   * Default property page url
-   *
-   * author Den@UD
-   */
-  static public function register_property_single_url()
-  {
-    global $wp_properties;
-    $properties = get_posts(array('post_type' => 'property'));
-    $post_id = $properties[0]->ID;
-    $post_url = get_permalink($post_id);
-
-    $wp_properties['configuration']['base_property_single_url'] = $post_url;
-    update_option('wpp_settings', $wp_properties);
-
-  }
-
-  /**
    * Perform WPP related things when a post is being deleted
    *
    * Makes sure all attached files and images get deleted.